Initial work on tightnening up the commands
