+++ source/commands/danger-pr.ts
@@ -1,7 +1,5 @@
 import * as program from "commander"
 import * as debug from "debug"
-import * as fs from "fs"
-import * as repl from "repl"
 import * as jsome from "jsome"
 
 import { FakeCI } from "../ci_source/providers/Fake"
@@ -10,15 +8,17 @@ import { GitHubAPI } from "../platforms/github/GitHubAPI"
 import { Executor } from "../runner/Executor"
 import { pullRequestParser } from "../platforms/github/pullRequestParser"
 import { runDangerfileEnvironment } from "../runner/DangerfileRunner"
-import { DangerContext } from "../runner/Dangerfile"
 import { dangerfilePath } from "./utils/file-utils"
+import validateDangerfileExists from "./utils/validateDangerfileExists"
+import openRepl from "./utils/repl"
 
 const d = debug("danger:pr")
 
 program
+  .usage("[options] <pr_url>")
+  .description("Emulate running Danger against an existing GitHub Pull Request.")
   .option("-v, --verbose", "Output more text to the stdout than a normal run")
   .option("-d, --dangerfile [filePath]", "Specify custom dangerfile other than default dangerfile.js")
-  .option("-r, --repl", "Drop into a Node REPL after evaluating the dangerfile")
   .parse(process.argv)
 
 const dangerFile = dangerfilePath(program)
@@ -46,23 +46,6 @@ if (program.args.length === 0) {
   }
 }
 
-function validateDangerfileExists(filePath: string): boolean {
-  let stat: fs.Stats | null = null
-  try {
-    stat = fs.statSync(filePath)
-  } catch (error) {
-    console.error(`Could not find a dangerfile at ${filePath}, not running against your PR.`)
-    process.exitCode = 1
-  }
-
-  if (!!stat && !stat.isFile()) {
-    console.error(`The resource at ${filePath} appears to not be a file, not running against your PR.`)
-    process.exitCode = 1
-  }
-
-  return !!stat && stat.isFile()
-}
-
 async function runDanger(source: FakeCI, platform: GitHub, file: string) {
   const config = {
     stdoutOnly: program.textOnly,
@@ -73,43 +56,9 @@ async function runDanger(source: FakeCI, platform: GitHub, file: string) {
 
   const runtimeEnv = await exec.setupDanger()
   const results = await runDangerfileEnvironment(file, undefined, runtimeEnv)
-  if (program["repl"]) {
+  if (program.repl) {
     openRepl(runtimeEnv.sandbox)
   } else {
     jsome(results)
   }
-}
-
-function openRepl(dangerContext: DangerContext): void {
-  /**
-   * Injects a read-only, global variable into the REPL
-   *
-   * @param {repl.REPLServer} repl The Node REPL created via `repl.start()`
-   * @param {string} name The name of the global variable
-   * @param {*} value The value of the global variable
-   */
-  function injectReadOnlyProperty(repl: repl.REPLServer, name: string, value: any) {
-    Object.defineProperty(repl["context"], name, {
-      configurable: false,
-      enumerable: true,
-      value,
-    })
-  }
-
-  /**
-   * Sets up the Danger REPL with `danger` and `results` global variables
-   *
-   * @param {repl.REPLServer} repl The Node REPL created via `repl.start()`
-   */
-  function setup(repl: repl.REPLServer) {
-    injectReadOnlyProperty(repl, "danger", dangerContext.danger)
-    injectReadOnlyProperty(repl, "results", dangerContext.results)
-  }
-
-  const dangerRepl = repl.start({ prompt: "> " })
-  setup(dangerRepl)
-  dangerRepl.on("exit", () => process.exit())
-  // Called when `.clear` is executed in the Node REPL
-  // This ensures that `danger` and `results` are not cleared from the REPL context
-  dangerRepl.on("reset", () => setup(dangerRepl))
 }
